+++ src/runtime/symbol_table.py
@@ -2,7 +2,7 @@
 # -*- coding:utf-8 -*-
 
 # Nougaro : a python-interpreted high-level programming language
-# Copyright (C) 2021-2023  Jean Dubois (https://github.com/jd-develop) <jd-dev@laposte.net>
+# Copyright (C) 2021-2024  Jean Dubois (https://github.com/jd-develop) <jd-dev@laposte.net>
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
@@ -11,7 +11,7 @@
 # __future__ imports (must be first)
 from __future__ import annotations
 # nougaro modules imports
-from src.constants import KEYWORDS
+from src.lexer.token_types import KEYWORDS
 # built-in python imports
 import pprint
 import difflib